requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
./pygraphviz-1.6-cp37-cp37m-win_amd64.whl
beautifulsoup4==4.9.3
certifi==2019.9.11
chardet==3.0.4
click==7.0
flask==1.0.2
future==0.18.1
graphviz==0.16
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
line-bot-sdk==1.14.0
markupsafe==1.1.1
numpy==1.19.4
pillow==8.0.1
pyimgur==0.6.0
python-dotenv==0.10.3
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==2.1; python_version >= '3.0'
transitions==0.6.9
urllib3==1.25.6
werkzeug==0.16.0
